+++ src/main/java/org/example/biomedbacktdd/dto/commands/EncryptMessageCommand.java
@@ -1,4 +1,4 @@
-package org.example.biomedbacktdd.DTO.commands;
+package org.example.biomedbacktdd.dto.commands;
 
 public class EncryptMessageCommand {
     private String url;
