+++ src/main/java/org/example/biomedbacktdd/services/DependentService.java
@@ -1,8 +1,8 @@
 package org.example.biomedbacktdd.services;
 
-import org.example.biomedbacktdd.DTO.commands.NewDependentCommand;
-import org.example.biomedbacktdd.DTO.results.NewDependentResult;
-import org.example.biomedbacktdd.DTO.viewmodels.NewDependentViewModel;
+import org.example.biomedbacktdd.dto.commands.NewDependentCommand;
+import org.example.biomedbacktdd.dto.results.NewDependentResult;
+import org.example.biomedbacktdd.dto.viewmodels.NewDependentViewModel;
 import org.example.biomedbacktdd.controllers.dependent.DependentController;
 import org.example.biomedbacktdd.entities.dependent.Dependent;
 import org.example.biomedbacktdd.exceptions.RequiredObjectIsNullException;
