+++ src/main/java/org/example/biomedbacktdd/services/EmailService.java
@@ -7,9 +7,9 @@ import com.sendgrid.SendGrid;
 import com.sendgrid.helpers.mail.Mail;
 import com.sendgrid.helpers.mail.objects.Content;
 import com.sendgrid.helpers.mail.objects.Email;
-import org.example.biomedbacktdd.DTO.commands.NewEmailCommand;
-import org.example.biomedbacktdd.DTO.results.NewEmailResult;
-import org.example.biomedbacktdd.DTO.viewmodels.NewEmailViewModel;
+import org.example.biomedbacktdd.dto.commands.NewEmailCommand;
+import org.example.biomedbacktdd.dto.results.NewEmailResult;
+import org.example.biomedbacktdd.dto.viewmodels.NewEmailViewModel;
 import org.example.biomedbacktdd.controllers.email.EmailController;
 import org.example.biomedbacktdd.entities.email.EmailHandler;
 import org.example.biomedbacktdd.exceptions.RequiredObjectIsNullException;
