+++ src/main/resources/application.properties
@@ -1,7 +1,5 @@
 spring.application.name=biomed-back-tdd
 
-spring.mvc.static-path-pattern=/swagger-ui/**
-
 spring.datasource.url=jdbc:postgresql://localhost:5432/zlotrackbandDB
 spring.datasource.username=postgres
 spring.datasource.password=admin123
@@ -9,17 +7,15 @@ spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.hibernate.ddl-auto=update
 
-sendgrid.api.key=
+cors.originPatterns=*
 
-# api.security.token.secret=${JWT_SECRET:my-secret-key}
-# security.jwt.token.secret-key=zloTrackBandSecretKey
-# security.jwt.token.secret-key.expire-length=3600000
+spring.mvc.static-path-pattern=/swagger-ui/**
 
 api.security.token.secret=${JWT_SECRET:my-secret-key}
 security.jwt.token.secret-key=
-security.jwt.token.secret-key.expire-length=60000
+security.jwt.token.secret-key.expire-length=86400000
 
-cors.originPatterns=*
+sendgrid.api.key=
 
 # Definir timezone para Sao Paulo / Brasil
 spring.jpa.properties.hibernate.jdbc.time_zone = America/Sao_Paulo
